Refactor/improve various aspects of type extraction.
In particulare, ensure that we can correctly handle named struct
definitions, including recursive structures.  To do this, we lift
the LLVM syntax extraction functions into a monad which keeps track
of the state of a depth-first search through the graph of LLVM types.

All types we encounter during extraction become entry points into the
graph, and we explore the collection of recursive definitions in a
depth-first manner as necessary, collecting definitions along the way.
These are eventually output into the module AST as type aliases.